backend/pyproject.toml
[project]
name = "fake-books"
version = "0.1.0"
description = ""
authors = [
    {name = "katsuyorii",email = "katsuyoriiii@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "email-validator (>=2.2.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
